requirements.txt
streamlit>=1.45.0
yfinance>=0.2.50
plotly>=6.0.0
pandas>=2.2.0
python-dotenv>=1.1.0

# Optional AI providers (install as needed)
# anthropic>=0.52.0
# openai>=1.82.0
